+++ script/cont_reader.h
@@ -6,7 +6,7 @@
 
 struct ContinuousGraspCandid
 {
-  std::vector< std::pair< std::pair<Eigen::Vector3d, Eigen::Vector3d>, Eigen::Quaterniond> > candids;
+  std::vector< std::pair< std::pair<Eigen::Vector3f, Eigen::Vector3f>, Eigen::Quaternionf> > candids;
 
   void loadConfig(std::string file_name)
   {
@@ -15,15 +15,15 @@ struct ContinuousGraspCandid
     yamlnode = YAML::LoadFile(file_name);
     for (int i=0 ;i< yamlnode["grasp_points"].size(); i++)
     {
-      std::pair< std::pair<Eigen::Vector3d, Eigen::Vector3d>, Eigen::Quaterniond> pose;
-      pose.first.first = Eigen::Vector3d::Map(yamlnode["grasp_points"][i]["upper_bound"].as<std::vector<double> >().data());
-      pose.first.second = Eigen::Vector3d::Map(yamlnode["grasp_points"][i]["lower_bound"].as<std::vector<double> >().data());
-      pose.second.coeffs() = Eigen::Vector4d::Map(yamlnode["grasp_points"][i]["orientation"].as<std::vector<double> >().data());
+      std::pair< std::pair<Eigen::Vector3f, Eigen::Vector3f>, Eigen::Quaternionf> pose;
+      pose.first.first = Eigen::Vector3f::Map(yamlnode["grasp_points"][i]["upper_bound"].as<std::vector<float> >().data());
+      pose.first.second = Eigen::Vector3f::Map(yamlnode["grasp_points"][i]["lower_bound"].as<std::vector<float> >().data());
+      pose.second.coeffs() = Eigen::Vector4f::Map(yamlnode["grasp_points"][i]["orientation"].as<std::vector<float> >().data());
       candids.push_back(pose);
     }
   }
 
-  Eigen::Isometry3d getGrasp(int index, double ratio)
+  Eigen::Isometry3f getGrasp(int index, float ratio)
   {
     if (index >= candids.size())
       throw std::out_of_range("index ");
@@ -33,7 +33,7 @@ struct ContinuousGraspCandid
 
     const auto & quat = candids[index].second;
 
-    Eigen::Isometry3d pose;
+    Eigen::Isometry3f pose;
     pose.translation() = (ub-lb) * ratio + lb;
     pose.linear() = quat.matrix();
